Fixes subtle bug in EditWordActivity.
How to reproduce: in EditWordActivity, click on the language button for
a translation entry.  When the "select language" dialog shows up, click
the Back button.  This dismisses the dialog, but the corresponding
translation entry still thinks it owns the dialog (i.e.
"hasLanguageDialogOpen" is still true).

Now click on the language button for the main word.  When the "select
language" dialog pops up, rotate the screen (such that the activity gets
restarted), and only then select a language.

TADA!  The language was changed, but for the translation entry, not for
the main word (which had opened the dialog before the rotation).

This was fixed by adding a listener on the "cancel" event for the
dialog, which now clears out the ownership from the corresponding
WordEntry.