Switches DB locks while exporting to RESERVED.
The EXCLUSIVE lock that was held before caused the application to crash
on screen rotation, because the underlying activity tried to get access
to the database in onCreate(), and it couldn't.

The RESERVED lock is still enough, since it's stronger than a SHARED
lock, which is required for holding while doing backups.