Updates TODO.
Gave up on trying to fix the layout for practice activity.  The problem
was that for long words, the language tag ended up outside the screen.
This was caused by the RelativeLayout which said that the word should be
centered on the screen, and the language tag should be right of it.
However, I can't think of a way to restrict the width of the word...

Switching to LinearLayout looks quite bad, so I'm giving up on this for
now.