Fixes small bug in PracticeActivity.
The ScoreRenderer was only reading the current time upon creation, but
was used repeatedly in PracticeActivity.  That lead to the situation
where words that expired after the practice session started would show
up for practice but would have their score rendered with the wrong color
(green, as in "not expired yet").