Fixes small detail in PracticeActivity.
The update of the database has been moved to when the activity view is
"off-screen" during the restart animation.  Like this the "I knew it" /
"I didn't know it" buttons have instant feedback, and if there are
delays accessing the database, they happen when the activity is not
visible.  This should theoretically make the "reload" feel more
realistic.