Keeps the "New Word" button visible when adding new words.
This refines the change added in dd358082c5... by keeping the "New Word"
button visible even when editing existing words.  Users found its
disappearance confusing and flow-disrupting.

On the implementation side, the intent action is now used for
determining the "mode": ACTION_INSERT is used when the user clicked "Add
New Words" in the MainActivity, and is kept even when editing existing
words in that context.  ACTION_EDIT is used from BrowseWordsActivity and
PracticeActivity (in this context, the "New Word" button is not shown).